requirements.txt
# Data generation and manipulation
faker==24.0.0
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2

# dbt
dbt-core==1.8.0
dbt-postgres==1.8.0

# Vector Store & Embeddings
langchain==0.1.16
langchain-community==0.0.34
httpx==0.27.2
openai==1.12.0
sentence-transformers==2.5.1
tiktoken==0.6.0

# Document Processing
pypdf==4.1.0
markdown==3.5.2

# MCP Server & FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
starlette==0.36.3

# Rate Limiting & Caching
slowapi==0.1.9
redis==5.0.1

# Monitoring & Logging
prometheus-client==0.20.0